requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
defusedxml==0.7.1
Django==4.2.9
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-compressor==4.4
django-rest-framework==0.1.0
django-sass-processor==1.4
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
libsass==0.22.0
pillow==10.2.0
pytz==2023.3.post1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
whitenoise==6.6.0
